[수정] - 리팩토링
