+++ README.md
@@ -20,17 +20,21 @@ SM 엔터테인먼트 사이트 주소 : https://smtownandstore.com/
 ### 참여한 팀원과 역할
 
 1. 황희윤 (메인 페이지 담당)
+
 - 헤더(페이지 상단의 메뉴 바), 메인 슬라이더, Pet for U (메인 페이지에서 보이는 상품 사진들) 구현
 
 2. 박태준 (상품 리스트 페이지, Footer)
+
 - 카테고리(동물 종류)별로 그에 해당하는 상품들의 목록을 보여주는 페이지 구현
 - 사이트 하단에 위치한 Footer 구현
 
 3. 이준혁 (상품 디테일 페이지, 장바구니 페이지)
+
 - 상품 개별 페이지를 구현하며 상품의 정보들을 디테일 하게 볼 수 있도록 구현
 - 상품을 구매하고 싶으면 로그인한 유저에 한해서 장바구니에 담을 수 있도록 구현
 
 4. 최은비 (로그인, 회원가입 페이지, 프로젝트 매니저 담당)
+
 - 유저가 특정한 정보를 입력하면 회원가입을 할 수 있게 회원가입 페이지 구현
 - 회원가입을 한 유저가 로그인을 하고 로그아웃도 할 수 있게 만듬
 - 프로젝트 전체적인 일정과 방향을 조율
@@ -43,13 +47,11 @@ SM 엔터테인먼트 사이트 주소 : https://smtownandstore.com/
 
 <img width="1433" alt="스크린샷 2022-01-07 오후 6 28 21" src="https://user-images.githubusercontent.com/87569550/148525591-67b13676-d2e1-4274-8a75-217a50855a69.png">
 
-
 #### 상품 리스트 페이지
 
 <img width="1440" alt="스크린샷 2022-01-07 오후 6 43 28" src="https://user-images.githubusercontent.com/87569550/148525659-10fce6cd-da16-448f-bdd2-28d410f0d43d.png">
 
 <img width="1438" alt="스크린샷 2022-01-07 오후 6 43 41" src="https://user-images.githubusercontent.com/87569550/148525668-10aca35e-4e76-48e6-9eff-45b0c36172bd.png">
-
 
 #### 상품 디테일 페이지
 
@@ -60,10 +62,6 @@ SM 엔터테인먼트 사이트 주소 : https://smtownandstore.com/
 
 <img width="1057" alt="스크린샷 2022-01-07 오후 6 52 51" src="https://user-images.githubusercontent.com/87569550/148526234-9354c67a-58aa-4c48-8627-b72c3f355c4a.png">
 
-
 #### 회원가입 페이지
 
 <img width="1042" alt="스크린샷 2022-01-07 오후 6 53 10" src="https://user-images.githubusercontent.com/87569550/148526244-6d903b13-be34-4112-98c0-c3c97c12c95a.png">
-
-
-
