+++ src/components/HeaderNav/HeaderNav.scss
@@ -1,8 +1,7 @@
 .HeaderNav {
   font-size: 16px;
-  position: relative;
+  position: fixed;
   z-index: 10000;
-  height: 9rem;
 
   .topBox {
     width: 100%;
