+++ src/pages/Cart/Cart.js
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
-import { Link } from "react-router-dom";
-import HeaderNav from "../../components/HeaderNav/HeaderNav";
-import Footer from "../../components/Footer/Footer";
+import { CartList } from "../ProductDetail/compo/CartList";
+import { TopBottom } from "../ProductDetail/compo/MiniCopo";
 import "./Cart.scss";
 
 const Cart = () => {
@@ -12,7 +11,7 @@ const Cart = () => {
 
   useEffect(() => {
     async function fetchData() {
-      await fetch(`http://localhost:8000/cart/list`, {
+      await fetch(`${process.env.REACT_APP_BASE_URL}/cart/list`, {
         method: "POST",
         mode: "cors",
         headers: {
@@ -34,8 +33,8 @@ const Cart = () => {
     }
     setTimeout(() => {
       fetchData();
-    }, 300);
-  }, [deleteOk]);
+    }, 100);
+  }, [deleteOk, userId]);
 
   const totalPrice = () => {
     let result = 0;
@@ -59,8 +58,12 @@ const Cart = () => {
     return result;
   };
 
+  const toComma = price => {
+    return price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+  };
+
   const deleteCart = async product_id => {
-    await fetch(`http://localhost:8000/cart`, {
+    await fetch(`${process.env.REACT_APP_BASE_URL}/cart`, {
       method: "delete",
       mode: "cors",
       headers: {
@@ -76,17 +79,16 @@ const Cart = () => {
   };
   return (
     <div className="Cart">
-      <HeaderNav />
       <div className="mainArea">
         <div className="title">Cart</div>
         <div className="product">
-          <div className="productField">
+          <section className="productField">
             <div>상품정보</div>
             <div>수량</div>
             <div>주문금액</div>
             <div>선택</div>
-          </div>
-          <div className="cartBox">
+          </section>
+          <section className="cartBox">
             {data.map((e, i) => {
               return (
                 <CartList
@@ -104,17 +106,14 @@ const Cart = () => {
                 />
               );
             })}
-          </div>
-          <div className="notice">
+          </section>
+          <section className="notice">
             장바구니에 담긴 상품은 30일 동안 보관됩니다.
-          </div>
-          <div className="totalArea">
+          </section>
+          <section className="totalArea">
             <div className="totalValue">
               <div>
-                {`₩ ` +
-                  totalPrice()
-                    .toString()
-                    .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}
+                {`₩ ` + toComma(totalPrice())}
                 <div className="allTotalPrice">총 상품금액</div>
               </div>
               <div>+</div>
@@ -123,146 +122,27 @@ const Cart = () => {
               </div>
               <div>-</div>
               <div>
-                {`₩ ` +
-                  salePrice()
-                    .toString()
-                    .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}
+                {`₩ ` + toComma(salePrice())}
                 <div className="salePrice">할인금액</div>
               </div>
               <div>=</div>
               <div style={{ color: "rgb(233, 52, 52)" }}>
-                {`₩ ` +
-                  (totalPrice() - salePrice())
-                    .toString()
-                    .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}{" "}
+                {`₩ ` + toComma(totalPrice() - salePrice())}
                 <div className="expectPrice" style={{ color: "black" }}>
                   결제예정금액
                 </div>
               </div>
             </div>
-          </div>
-          <div className="btnBox">
+          </section>
+          <section className="btnBox">
             <button className="orderBtn">전체상품 주문</button>
             <button>선택상품 주문</button>
             <button>🛍️ 선택상품 선물</button>
-          </div>
-          <div className="topBottom">
-            <span
-              className="top"
-              onClick={() => {
-                window.scrollTo({ top: 0, behavior: "smooth" });
-              }}
-            >
-              <img src="/images/totop.png" alt="top" />
-            </span>
-            <span
-              className="bottom"
-              onClick={() => {
-                window.scrollTo({ top: 10000, behavior: "smooth" });
-              }}
-            >
-              <img src="/images/totop.png" alt="bottom" />
-            </span>
-          </div>
+          </section>
+          <TopBottom />
         </div>
       </div>
-      <Footer />
     </div>
   );
 };
-
-const CartList = ({
-  img,
-  name,
-  price,
-  sale,
-  deleteCart,
-  product_id,
-  cate,
-  productQuantity,
-  setProductQuantity,
-  index,
-}) => {
-  // const [productQuantity, setProductQuantity] = useState(quantity);
-
-  return (
-    <div className="productDetail">
-      <div className="dot">◾</div>
-      <div className="productImg">
-        <Link to={`/category/${cate}/product/${product_id}`}>
-          <img src={img} alt="이미지" />
-        </Link>
-      </div>
-      <div className="productName">
-        <div className="cateName">{cate.toUpperCase()}</div>
-        <div className="name">{name.toUpperCase()}</div>
-      </div>
-
-      <div className="productQuantity">
-        <button
-          className="minus"
-          onClick={() => {
-            const copy = [...productQuantity];
-            copy[index] -= 1;
-            productQuantity[index] <= 1
-              ? alert("최소 주문수량은 1개 입니다.")
-              : setProductQuantity(copy);
-          }}
-        >
-          -
-        </button>
-        {productQuantity[index]}
-        <button
-          className="plus"
-          onClick={() => {
-            const copy = [...productQuantity];
-            copy[index] += 1;
-            productQuantity[index] >= 10
-              ? alert("최대 주문수량은 10개 입니다.")
-              : setProductQuantity(copy);
-          }}
-        >
-          +
-        </button>
-      </div>
-      <div className="productPrice" style={{ fontWeight: "bold" }}>
-        {sale ? (
-          <Price price={price} sale={sale} quantity={productQuantity[index]} />
-        ) : (
-          `₩ ` +
-          Math.round(price * productQuantity[index])
-            .toString()
-            .replace(/\B(?=(\d{3})+(?!\d))/g, ",")
-        )}
-      </div>
-      <div className="prodictSelect">
-        <button
-          onClick={() => {
-            deleteCart(product_id);
-          }}
-        >
-          삭제
-        </button>
-      </div>
-    </div>
-  );
-};
-
-const Price = ({ sale, price, quantity }) => {
-  return (
-    <div className="priceBox">
-      <span className="afterPrice">
-        {`₩ ` +
-          Math.round(price * (1 - sale / 100) * quantity)
-            .toString()
-            .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}
-      </span>
-      <span className="beforePrice">
-        {(price * quantity).toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}
-      </span>
-      <span className="saleRate">{sale}%</span>
-    </div>
-  );
-};
-
 export default Cart;
